server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
github-copilot-sdk>=0.1.32
python-dotenv>=1.0.0
